Unite key and buttons voting in `VoteButton` component
