+++ packages/webapp/src/App.tsx
@@ -25,8 +25,7 @@ import { setActiveAccount } from './state/slices/account';
 import { openVoteSocket, markVoterInactive } from './middleware/voteWebsocket';
 import { openEthereumSocket } from './middleware/alchemyWebsocket';
 import NotificationToast from './components/NotificationToast';
-import VoteWithKeys from './components/VoteWithKeys';
-
+import CornerHelpText from './components/CornerHelpText';
 
 
 function App() {
@@ -78,7 +77,7 @@ function App() {
       <Banner />
       <Documentation />
       <Footer/>
-      <VoteWithKeys />
+      <CornerHelpText />
       <NotificationToast />
     </div>
   );
