+++ packages/webapp/src/components/VoteButton/index.tsx
@@ -3,6 +3,7 @@ import classes from './VoteButton.module.css';
 import { VOTE_OPTIONS, setCurrentVote } from '../../state/slices/vote';
 import { useAppDispatch, useAppSelector } from '../../hooks';
 import { sendVote } from '../../middleware/voteWebsocket';
+import useEventListener from '@use-it/event-listener';
 
 export enum EMOJI_TYPE {
     dislike = '👎',
@@ -14,6 +15,12 @@ const voteToEmoji: Record<VOTE_OPTIONS, string> = {
   [VOTE_OPTIONS['voteDislike']]: '👎',
   [VOTE_OPTIONS['voteShrug']]: '🤷‍♂️',
   [VOTE_OPTIONS['voteLike']]: '👍'
+};
+
+const voteToKey: Record<VOTE_OPTIONS, string> = {
+  [VOTE_OPTIONS.voteDislike]: 'ArrowLeft',
+  [VOTE_OPTIONS.voteShrug]: '-100',
+  [VOTE_OPTIONS.voteLike]: 'ArrowRight'
 };
 
 const VoteButton: React.FC<{voteType: VOTE_OPTIONS}> = props => {
@@ -30,7 +37,8 @@ const VoteButton: React.FC<{voteType: VOTE_OPTIONS}> = props => {
   const { voteType } = props;
   const voteNotSelected = (currentVote !== undefined) && currentVote !== voteType;
   const dispatch = useAppDispatch();
-  const changeVote = () => {
+  
+  const changeVote = (voteType: VOTE_OPTIONS) => {
     if (currentVote || !wsConnected) return;
     
     dispatch(setCurrentVote(voteType));
@@ -39,11 +47,20 @@ const VoteButton: React.FC<{voteType: VOTE_OPTIONS}> = props => {
 
   const disabled = voteNotSelected || (!votingActive || activeAuction) || blockHash !== votingBlockHash
 
+  useEventListener("keydown", (key) => {
+    const event = key as KeyboardEvent
+
+    if (event.key === voteToKey[voteType] && !disabled) {
+      changeVote(voteType);
+    }
+  });
+
   return (
-      <button className={currentVote === voteType ? clsx(classes.voteButton, classes.selected) : classes.voteButton} onClick={changeVote}
-      disabled={disabled}>
-        <span className={classes.voteEmojiText}> {voteToEmoji[voteType]} </span>
-        <span className={classes.voteText}> {voteCounts[voteType]} </span>
+      <button className={currentVote === voteType ? clsx(classes.voteButton, classes.selected) : classes.voteButton}
+        onClick={() => changeVote(voteType)}
+        disabled={disabled}>
+          <span className={classes.voteEmojiText}> {voteToEmoji[voteType]} </span>
+          <span className={classes.voteText}> {voteCounts[voteType]} </span>
       </button>
   );
 };
